Cementerios_Mapas

+++ AtoGobMx/Models/Cementerio.cs
@@ -15,8 +15,8 @@ namespace AtoGobMx.Models
         public string Localidad { get; set; }
         public string Calle { get; set; }
         public int NumeroExterior { get; set; }
-        public float Longitud { get; set; }
-        public float Latitud { get; set; }
+        public decimal Longitud { get; set; }
+        public decimal Latitud { get; set; }
         public bool Archivado { get; set; }
 
         [JsonIgnore]
